Update readme
Was missing some information. Added service provider/facade/config (which might be obvious to some).
Also, the instance example couldn't be used, because it requires the app container in construct, not the key/secret.
And the auth example also didn't work. Can you pass an array of routes? I didn't think so.
Made the post/get callback to the same url.
Also, the `Flickering::getOpauth();` didn't return a response object, but just sets the headers (I think), so an exception was shown briefly, before being redirected.